refactor(common): rename `types` to `schema`
