+++ renderer/hooks/useGameList.ts
@@ -1,6 +1,6 @@
 import { createSignal } from 'solid-js';
 
-import { GameList } from '../../common/types';
+import { GameList } from '../../common/schema';
 
 const useGameList = () => {
   const [gameList, setGameList] = createSignal<GameList>({});
