+++ renderer/settings/containers/GeneralContainer.tsx
@@ -7,7 +7,7 @@ import Selector from '../../components/Select';
 
 import useConfig from '../../hooks/useConfig';
 import { getTranslation } from '../../../common/intl';
-import { Config } from '../../../common/types';
+import { Config } from '../../../common/schema';
 import Modal from '../../components/Modal';
 import Switch from '../../components/Switch';
