+++ src/config/lyric-mapper.ts
@@ -4,7 +4,7 @@ import { State } from './state';
 
 import { defaultConfigDirectory } from './config';
 
-import { LyricMapper, LyricMapperSchema } from '../../common/types';
+import { LyricMapper, LyricMapperSchema } from '../../common/schema';
 
 
 const lyricPath = path.join(defaultConfigDirectory, 'lyrics.json');
