+++ src/config/theme-list.ts
@@ -7,7 +7,7 @@ import { State } from './state';
 import { defaultConfigDirectory } from './config';
 
 import { DEFAULT_STYLE } from '../../common/constants';
-import { StyleConfig, ThemeList } from '../../common/types';
+import { StyleConfig, ThemeList } from '../../common/schema';
 
 let throttleTimer: NodeJS.Timeout | null = null;
 export const themeList = new State<ThemeList>({});
